# swagger.json
{"openapi":"3.1.0","info":{"title":"Authentication service","version":"0.1.0"},"paths":{"/auth/register":{"post":{"tags":["Authentication"],"summary":"User Registration","description":"Register a new user.\n\n- **username**: The user username.\n- **email**: The user email (unique).\n- **password**: The user password.\n- **user_type**: 0 for normal users, 1 for couriers.\n\nReturns the created user.","operationId":"register_user_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/verify-email":{"post":{"tags":["Authentication"],"summary":"Verify User Email","description":"Verify the user email with a verification code that's send to the email.\n\n- **email**: The email to verify.\n- **code**: The verification code sent via email.\n\nReturns success message if the email can be verified.","operationId":"verify_email_auth_verify_email_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"User Login","description":"Authenticate a user and return a temporary access token.\n\n- **email**: The user registered email.\n- **password**: The user password.\n\nReturns a JWT access token if authentication is successful.","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/user/{username}":{"get":{"tags":["User Information"],"summary":"Get User Info","description":"Retrieve user information by username.\n\n- **username**: The username of the user.\n\nReturns the user's public information.","operationId":"get_user_info_auth_user__username__get","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"user_type":{"type":"integer","maximum":1.0,"minimum":0.0,"title":"User Type","default":0}},"type":"object","required":["username","email","password"],"title":"UserCreate"},"UserModel":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"hashed_password":{"type":"string","title":"Hashed Password"},"verification_code":{"type":"string","title":"Verification Code"},"is_verified":{"type":"boolean","title":"Is Verified","default":false},"user_type":{"type":"integer","maximum":1.0,"minimum":0.0,"title":"User Type","default":0},"created_at":{"type":"string","format":"date-time","title":"Created At","default":"2025-03-18T19:29:54.367599"}},"type":"object","required":["username","email","hashed_password","verification_code"],"title":"UserModel"},"UserResponse":{"properties":{"id":{"type":"string","title":"Id"},"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"is_verified":{"type":"boolean","title":"Is Verified"},"user_type":{"type":"integer","title":"User Type"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","username","email","is_verified","user_type","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerifyEmailRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"code":{"type":"string","title":"Code"}},"type":"object","required":["email","code"],"title":"VerifyEmailRequest"}}}}